+++ providers/common/ciphers/block.c
@@ -7,11 +7,9 @@
  * https://www.openssl.org/source/license.html
  */
 
-#include <string.h>
-#include <openssl/evp.h>
-#include <openssl/err.h>
-#include "ciphers_locl.h"
 #include <assert.h>
+#include <openssl/err.h>
+#include "cipher_locl.h"
 #include "internal/providercommonerr.h"
 
 /*
